requirements.txt
numpy>=1.14.0
pandas>=0.23.4
gdown>=3.10.1
tqdm>=4.30.0
Pillow>=5.2.0
opencv-python>=4.2.0.34
opencv-contrib-python>=4.3.0.36
tensorflow>=1.9.0
keras>=2.2.0
Flask>=1.1.2
mtcnn>=0.1.0
# lightgbm>=2.3.1
# dlib>=19.20.0
retina-face>=0.0.1
elasticsearch==7.10.0
cmake
pyyaml